db: store optional JSON column with data series
The current DatabaseULog.save() and DatabaseULog.load() methods both use
the BLOB column ValueArray to store the packed data from the data series
of a data set. This is space and time efficient, but has the issue that
we cannot access the data series directly in sqlite without processing
it in an external program (e.g. python).

Accessing the data series as rows allows for interesting use cases like
integration with plotting software (e.g. Grafana) and for fast
calculation of aggregate fields such as averages, sums and moving
averages.

This commit solves this problem by adding an optional flag "append_json"
to the DatabaseULog.save() function. When set to true, another column
ValueJson of sqlite type JSON will be populated in addition to the
aforementioned ValueArray BLOB. Using the sqlite function json_each,
the data in the ValueJson column can be expanded into rows for further
processing.

This process is not as efficient in time as the BLOB process, so load()
will always use the BLOB. Furthermore, storing all the ULog data in JSON
strings is obviously not as efficient as packed byte arrays, so the size
of the database will grow by some factor when enabled (the author
guesses it not to exceed 10x).